Moving Production Log DTOs into their own Namespace
