+++ MESS/MESS.Data/DTO/ProductionLogDTOs/ProductionLogStepDTO.cs
@@ -1,6 +1,4 @@
-﻿using MESS.Data.Models;
-
-namespace MESS.Data.DTO;
+﻿namespace MESS.Data.DTO.ProductionLogDTOs;
 
 /// <summary>
 /// Represents a step in the production log. Used primarily for client-side caching purposes.
@@ -11,12 +9,12 @@ public class ProductionLogStepDTO
     /// <summary>
     /// Gets or sets the ID of the work instruction step.
     /// </summary>
-    public int WorkInstructionStepId { get; set; }
+    public int WorkInstructionStepId { get; init; }
 
     /// <summary>
     /// Gets or sets the ID of the production log.
     /// </summary>
-    public int ProductionLogId { get; set; }
+    public int ProductionLogId { get; init; }
 
     /// <summary>
     /// Gets or sets the collection of attempts for this step.
