+++ MESS/MESS.Services/LocalCacheManager/ILocalCacheManager.cs
@@ -1,4 +1,5 @@
 ﻿using MESS.Data.DTO;
+using MESS.Data.DTO.ProductionLogDTOs;
 
 namespace MESS.Services.BrowserCacheManager;
 using Data.Models;
