+++ MESS/MESS.Services/LocalCacheManager/LocalCacheManager.cs
@@ -1,4 +1,5 @@
 ﻿using MESS.Data.DTO;
+using MESS.Data.DTO.ProductionLogDTOs;
 using MESS.Services.BrowserCacheManager;
 using Microsoft.AspNetCore.Components.Server.ProtectedBrowserStorage;
 using Serilog;
