+++ MESS/MESS.Tests/UI_Testing/ProductionLog/ProductionLogCreationTests.cs
@@ -2,7 +2,7 @@
 using Bunit;
 using Bunit.TestDoubles;
 using MESS.Blazor.Components.Pages.ProductionLog;
-using MESS.Data.DTO;
+using MESS.Data.DTO.ProductionLogDTOs;
 using MESS.Services.ApplicationUser;
 using MESS.Services.BrowserCacheManager;
 using MESS.Services.Product;
